requirements.txt
Django==2.2.3
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-js-asset==1.2.2
djangorestframework==3.10.2
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.3
